extracting abstract IntentModel and creating DefaultIntentModel
