+++ RankSys-diversity/src/main/java/es/uam/eps/ir/ranksys/diversity/intentaware/metrics/ERRIA.java
@@ -66,7 +66,7 @@ public class ERRIA<U, I, F> extends AbstractRecommendationMetric<U, I> {
     @Override
     public double evaluate(Recommendation<U, I> recommendation) {
         RelevanceModel.UserRelevanceModel<U, I> userRelModel = relModel.getModel(recommendation.getUser());
-        IntentModel<U, I, F>.AbstractUserIntentModel uim = intentModel.getModel(recommendation.getUser());
+        IntentModel<U, I, F>.UserIntentModel uim = intentModel.getModel(recommendation.getUser());
 
         double[] erria = {0.0};
